requirements.txt
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
DateTime==4.3
dnspython==1.16.0
email-validator==1.1.1
flake8==3.8.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
google-api-core==1.20.1
google-auth==1.17.2
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pyrsistent==0.16.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.1
urllib3==1.25.9
validate-docbr==1.7.0
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.0
uwsgi==2.0.19.1